Improved the board printed to the CLI to include the coordenates, making it easier to see the coords where you want to play
changed the win condition to return 1 or 2 values possible so we can define which player won.

created inputCoords function which runs the steps and validations to read the inputs from user, save the coords in an array and return a pointer to the array

cleaned up code to remove redundant code, such as excessive cout commands

simplified the logic to print which player is playing reducing the amount of lines of code to half on that feature.

used the point to array from inputCoords to read and validate if the coord the player selected are valid. If a player had already played in those coordenated, it is asked to input valid ones.